Pull app components into separate file
